+++ lint-janitor.config.js
@@ -2,6 +2,7 @@ module.exports = {
   esFiles: ['test/**/*.js', 'test/**/*.ts', 'src/**/*.ts'],
   styleFiles: ['test/**/*.less', 'test/**/*.scss'],
   htmlFiles: ['test/**/*.html'],
+  markdownFiles: ['./*.md'],
   jsonFiles: ['*.json', '.eslintrc', '.htmlhintrc', 'lint/**/*.json', 'lint/.eslintrc', 'lint/.htmlhintrc', 'test/*.json', 'src/**/*.json'],
   yamlFiles: ['./.travis.yml', './.sass-lint.yml', './lint/.sass-lint.yml']
 };