+++ src/index.ts
@@ -12,5 +12,6 @@ export { ZJsonLint } from './json-lint/json-lint.class';
 export { IZLintJanitorArgs } from './lint-janitor/lint-janitor-args.interface';
 export { IZLintJanitorOptions } from './lint-janitor/lint-janitor-options.interface';
 export { ZLintJanitor } from './lint-janitor/lint-janitor.class';
+export { ZMarkdownLint } from './markdown-lint/markdown-lint.class';
 export { ZSilentLint } from './silent-lint/silent-lint.class';
 export { ZYamlLint } from './yaml-lint/yaml-lint.class';