+++ src/lint-janitor/lint-janitor-options.interface.ts
@@ -2,10 +2,12 @@ export interface IZLintJanitorOptions {
   esConfig?: string;
   styleConfig?: string;
   htmlConfig?: string;
+  markdownConfig?: string;
 
   esFiles?: string[];
   styleFiles?: string[];
   htmlFiles?: string[];
+  markdownFiles?: string[];
   jsonFiles?: string[];
   yamlFiles?: string[];
 }